examples/streamlit/postgres-connection/requirements.txt
streamlit
SQLAlchemy
datasets
psycopg2-binary
